vs: fix a few compiler warnings

+++ pcsx2/DEV9/pcap_io.cpp
@@ -298,7 +298,7 @@ int pcap_io_recv(void* packet, int max_len)
 
 	if ((pcap_next_ex(adhandle, &header, &pkt_data1)) > 0)
 	{
-		if (header->len > max_len)
+		if ((int)header->len > max_len)
 			return -1;
 
 		memcpy(packet, pkt_data1, header->len);
