+++ pcsx2/Gif.cpp
@@ -140,7 +140,7 @@ int GIF_Fifo::read_fifo()
 
 	GIF_LOG("GIF FIFO Read %d QW from FIFO Current Size %d", sizeRead, fifoSize);
 
-	if (sizeRead < fifoSize)
+	if (sizeRead < (int)fifoSize)
 	{
 		if (sizeRead > 0)
 		{
